refactor: 날짜 util 함수 분리 #5
